+++ KOF.Core/Services/GameService.cs
@@ -169,7 +169,7 @@ public partial class GameService
         else if (commandType == 1)
         {
             ushort count = msg.Read<ushort>();
-            short[] playerIds = new short[count];
+            int[] playerIds = new int[count];
             msg.Read(playerIds.AsSpan());
 
             var playerIdList = playerIds.ToList();
@@ -196,7 +196,7 @@ public partial class GameService
     [MessageHandler(MessageID.WIZ_MOVE)]
     public Task MsgRecv_PlayerMovement(Session session, Message msg)
     {
-        short socketId = msg.Read<short>();
+        int socketId = msg.Read<int>();
 
         float will_x = msg.Read<ushort>() / 10.0f;
         float will_y = msg.Read<ushort>() / 10.0f;
@@ -246,7 +246,7 @@ public partial class GameService
         {
             if (!session.Client.CharacterHandler.PlayerList.Any(x => x.Id == socketId)) //Ghost Player?
             {
-                short[] playerIds = new short[1] { socketId };
+                int[] playerIds = new int[1] { socketId };
                 return session.SendAsync(MessageBuilder.MsgSend_UserRequest(1, playerIds));
             }
             else
@@ -317,7 +317,7 @@ public partial class GameService
                 break;
 
             case InOutType.INOUT_OUT:
-                ushort npcId = msg.Read<ushort>();
+                int npcId = msg.Read<int>();
 
                 lock (session.Client.CharacterHandler.NpcList)
                     session.Client.CharacterHandler.NpcList.RemoveAll((x) => x.Id == npcId);
@@ -344,7 +344,7 @@ public partial class GameService
             return Task.CompletedTask;
         }
 
-        short[] npcIds = new short[count];
+        int[] npcIds = new int[count];
         msg.Read(npcIds.AsSpan());
 
         var npcIdList = npcIds.ToList();
@@ -377,7 +377,7 @@ public partial class GameService
 
         if (commandType == 1 || commandType == 2)
         {
-            short npcId = msg.Read<short>();
+            int npcId = msg.Read<int>();
             float positionX = msg.Read<ushort>() / 10.0f;
             float positionY = msg.Read<ushort>() / 10.0f;
             float positionZ = msg.Read<ushort>() / 10.0f;
@@ -386,7 +386,7 @@ public partial class GameService
 
             if (!session.Client.CharacterHandler.NpcList.Any(x => x.Id == npcId)) //Ghost Monster?
             {
-                short[] npcIds = new short[1] { npcId };
+                int[] npcIds = new int[1] { npcId };
                 return session.SendAsync(MessageBuilder.MsgSend_NpcRequest(1, npcIds));
             }
             else
@@ -411,7 +411,7 @@ public partial class GameService
     {
         var character = session.Client.Character;
 
-        ushort targetId = msg.Read<ushort>();
+        int targetId = msg.Read<int>();
 
         if (character.Id == targetId)
         {
@@ -530,7 +530,7 @@ public partial class GameService
     [MessageHandler(MessageID.WIZ_TARGET_HP)]
     public Task MsgRecv_TargetHp(Session session, Message msg)
     {
-        short targetId = msg.Read<short>();
+        int targetId = msg.Read<int>();
         byte byUpdateImmediately = msg.Read<byte>();
         uint iTargetHPMax = msg.Read<uint>();
         uint iTargetHPCur = msg.Read<uint>();
@@ -601,8 +601,8 @@ public partial class GameService
 
             case AttackResult.ATTACK_TARGET_DEAD:
             case AttackResult.ATTACK_TARGET_DEAD_OK:
-                _ = msg.Read<ushort>(); // attackerId
-                var targetId = msg.Read<ushort>();
+                _ = msg.Read<int>(); // attackerId
+                var targetId = msg.Read<int>();
 
                 if (targetId == session.Client.Character.Id)
                 {
@@ -657,8 +657,8 @@ public partial class GameService
             case SkillMagicType.SKILL_MAGIC_TYPE_CASTING:
                 {
                     var skillId = msg.Read<int>();
-                    var sourceId = msg.Read<short>();
-                    var targetId = msg.Read<short>();
+                    var sourceId = msg.Read<int>();
+                    var targetId = msg.Read<int>();
 
                     session.Client.CharacterHandler.SkillCastingProcess(skillId, sourceId, targetId);
                 }
@@ -667,8 +667,8 @@ public partial class GameService
             case SkillMagicType.SKILL_MAGIC_TYPE_FLYING:
                 {
                     var skillId = msg.Read<int>();
-                    var sourceId = msg.Read<short>();
-                    var targetId = msg.Read<short>();
+                    var sourceId = msg.Read<int>();
+                    var targetId = msg.Read<int>();
 
                     session.Client.CharacterHandler.SkillFlyingProcess(skillId, sourceId, targetId);
                 }
@@ -677,8 +677,8 @@ public partial class GameService
             case SkillMagicType.SKILL_MAGIC_TYPE_EFFECTING:
                 {
                     var skillId = msg.Read<int>();
-                    var sourceId = msg.Read<short>();
-                    var targetId = msg.Read<short>();
+                    var sourceId = msg.Read<int>();
+                    var targetId = msg.Read<int>();
 
                     session.Client.CharacterHandler.SkillEffectingProcess(skillId, sourceId, targetId);
                 }
@@ -687,8 +687,8 @@ public partial class GameService
             case SkillMagicType.SKILL_MAGIC_TYPE_FAIL:
                 {
                     var skillId = msg.Read<int>();
-                    var sourceId = msg.Read<short>();
-                    var targetId = msg.Read<short>();
+                    var sourceId = msg.Read<int>();
+                    var targetId = msg.Read<int>();
 
                     List<short> data = new();
 
@@ -874,7 +874,7 @@ public partial class GameService
     [MessageHandler(MessageID.WIZ_ITEM_DROP)]
     public Task MsgRecv_ItemBundleDrop(Session session, Message msg)
     {
-        var npcId = msg.Read<short>();
+        var npcId = msg.Read<int>();
         var bundleId = msg.Read<uint>();
         var itemCount = msg.Read<byte>();
 
@@ -1012,7 +1012,7 @@ public partial class GameService
     [MessageHandler(MessageID.WIZ_LEVEL_CHANGE)]
     public Task MsgRecv_MyInfo_LevelChange(Session session, Message msg)
     {
-        var id = msg.Read<ushort>();
+        var id = msg.Read<int>();
         var level = msg.Read<byte>();
 
         if (id == session.Client.Character.Id)
@@ -1109,7 +1109,7 @@ public partial class GameService
 
             case PartyUpdateType.Insert:
                 {
-                    ushort insertMemberId = msg.Read<ushort>();
+                    int insertMemberId = msg.Read<int>();
                     string invitedName = msg.Read(true, "gb2312");
 
                     var client = ClientHandler.ClientList.FirstOrDefault(x => x != null && x.Name == invitedName)!;
@@ -1196,7 +1196,7 @@ public partial class GameService
 
             case PartyUpdateType.Leave:
                 {
-                    var leftMemberId = msg.Read<ushort>();
+                    var leftMemberId = msg.Read<int>();
 
                     lock (session.Client.Character.Party.Members)
                         session.Client.Character.Party.Members.RemoveAll((partyMember) => partyMember.MemberId == leftMemberId);
@@ -1230,7 +1230,7 @@ public partial class GameService
 
             case PartyUpdateType.HealthManaChange:
                 {
-                    var HPMPMemberId = msg.Read<ushort>();
+                    var HPMPMemberId = msg.Read<int>();
 
                     session.Client.Character.Party.Members.ForEach((partyMember) =>
                     {
@@ -1248,7 +1248,7 @@ public partial class GameService
 
             case PartyUpdateType.LevelChange:
                 {
-                    var levelMemberId = msg.Read<ushort>();
+                    var levelMemberId = msg.Read<int>();
 
                     session.Client.Character.Party.Members.ForEach((partyMember) =>
                     {
@@ -1261,7 +1261,7 @@ public partial class GameService
 
             case PartyUpdateType.ClassChange:
                 {
-                    var classMemberId = msg.Read<ushort>();
+                    var classMemberId = msg.Read<int>();
 
                     session.Client.Character.Party.Members.ForEach((partyMember) =>
                     {
@@ -1275,7 +1275,7 @@ public partial class GameService
 
             case PartyUpdateType.StatusChange:
                 {
-                    var statusChangeMemberId = msg.Read<ushort>();
+                    var statusChangeMemberId = msg.Read<int>();
 
                     session.Client.Character.Party.Members.ForEach((partyMember) =>
                     {
@@ -1299,7 +1299,7 @@ public partial class GameService
     [MessageHandler(MessageID.WIZ_STATE_CHANGE)]
     public Task MsgRecv_StateChange(Session session, Message msg)
     {
-        ushort socketId = msg.Read<ushort>();
+        int socketId = msg.Read<int>();
         byte subPacket = msg.Read<byte>();
         uint state = msg.Read<uint>();
 
@@ -1474,7 +1474,7 @@ public partial class GameService
         {
             case (byte)TradeSubPacket.TRADE_REQ:
                 {
-                    var otherId = msg.Read<short>();
+                    var otherId = msg.Read<int>();
                     session.Client.CharacterHandler.ExhangeRequestProcess(otherId);
                 }
                 break;
@@ -1754,7 +1754,7 @@ public partial class GameService
     [MessageHandler(MessageID.WIZ_ROTATE)]
     public Task MsgRecv_Rotation(Session session, Message msg)
     {
-        var id = msg.Read<short>();
+        var id = msg.Read<int>();
         var rotate = msg.Read<short>() / 100.0f;
 
         session.Client.Character.Rotation = rotate;
